KAN-93 refactor: ImagePosts
