--- src/common/components/Posts/model/postsType.ts
@@ -1,30 +1,0 @@
-export type ImagePostResponse = {
-  createdAt: string
-  fileSize: number
-  height: number
-  id: number
-  uploadId: string
-  url: string
-  width: number
-}
-
-export type Owner = {
-  firstName: string
-  lastName: string
-}
-
-export type Post = {
-  avatarOwner: string
-  avatarWhoLikes: boolean
-  createdAt: string
-  description: string
-  id: number
-  images: ImagePostResponse[]
-  isLiked: boolean
-  likesCount: number
-  location: string
-  owner: Owner
-  ownerId: number
-  updatedAt: string
-  userName: string
-}
